[REEF-268] Moved `ConfigurationProvider` into Tang
This moves the `ConfigurationProvider` interface into Tang such that
e.g. Wake can use it.

JIRA:
  [REEF-268](https://issues.apache.org/jira/browse/REEF-268)

Pull Request:
  This closes #
